Extract `normalizePath` function into own module
I need this function to normalize paths coming from watchman
subscription events. That change will come in a later commit.
