+++ lib/Configuration.js
@@ -2,12 +2,12 @@
 
 import path from 'path';
 
-import escapeRegExp from 'lodash.escaperegexp';
 import minimatch from 'minimatch';
 import semver from 'semver';
 
 import FileUtils from './FileUtils';
 import JsModule from './JsModule';
+import normalizePath from './normalizePath';
 import requireResolve from './requireResolve';
 import version from './version';
 
@@ -144,17 +144,6 @@ function checkForUnknownConfiguration(config: Object): Array<string> {
   });
 
   return messages;
-}
-
-function normalizePath(rawPath: string): string {
-  if (!rawPath) {
-    return './';
-  }
-  const normalized = rawPath.replace(RegExp(`^${escapeRegExp(process.cwd())}`), '.');
-  if (normalized.startsWith('.')) {
-    return normalized;
-  }
-  return `./${normalized}`;
 }
 
 /**
